+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+- unit tests for filter input data types
+
 ### Changed
 - factory method for filter inupt data types from `createWhatEverType` to `fromUserInput`